backend/pyproject.toml
[project]
name = "aichat"
version = "0.1.0"
description = "JeikChat 智能客服系统 - 后端"
requires-python = ">=3.11, <3.14"

dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy",
    "pydantic",
    "pydantic-settings",
    "python-multipart",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "langchain",
    "langchain-community",
    "langchain-openai",
    "chromadb",
    "sentence-transformers",
    "pypdf",
    "python-docx",
    "openpyxl",
    "pandas",
    "python-magic",
    "aiofiles",
    "httpx",
    "tenacity",
    "langgraph-checkpoint-postgres",
    "psycopg[binary,pool]",
    "langchain-deepseek>=1.0.1",
    "asyncio>=4.0.0",
    "dict>=2018.11.19",
    "any>=0.0.2",
    "node>=2.0.0",
    "langchain-google-genai>=4.2.1",
    "unstructured[all-docs]>=0.21.5",
    "langchain-chroma>=1.1.0",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-unstructured>=1.0.1",
    "libmagic>=1.0",
    "langchain-qdrant>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
]

[project.scripts]
jeikchat = "cli:main"

[tool.setuptools]
py-modules = ["cli"]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"
